Move TSV format code to static class
